refactor: extract en/decrypt_file
