+++ crates/bestool/src/actions/crypto.rs
@@ -1,13 +1,10 @@
-use std::io::{stderr, IsTerminal as _};
-
 use clap::{Parser, Subcommand};
-use indicatif::{ProgressBar, ProgressBarIter, ProgressStyle};
 use miette::Result;
-use tokio::io::AsyncRead;
 
 use super::Context;
 
-pub mod key;
+pub mod keys;
+pub mod streams;
 
 /// Cryptographic operations.
 #[derive(Debug, Clone, Parser)]
@@ -26,22 +23,4 @@ super::subcommands! {
 	encrypt => Encrypt(EncryptArgs),
 	hash => Hash(HashArgs),
 	keygen => Keygen(KeygenArgs)
-}
-
-/// Wraps a [`tokio::io::AsyncRead`] with an [`indicatif::ProgressBar`].
-///
-/// The progress bar outputs to stderr iff that's terminal, and nothing is displayed otherwise.
-pub(crate) fn with_progress_bar<R: AsyncRead + Unpin>(
-	expected_length: u64,
-	reader: R,
-) -> ProgressBarIter<R> {
-	if stderr().is_terminal() {
-		let style = ProgressStyle::default_bar()
-			.template("[{bar:.green/blue}] {wide_msg} {binary_bytes}/{binary_total_bytes} ({eta})")
-			.expect("BUG: progress bar template invalid");
-		ProgressBar::new(expected_length).with_style(style)
-	} else {
-		ProgressBar::hidden()
-	}
-	.wrap_async_read(reader)
 }
